Fix and clean EloqModule. (#65)
* CheckAndUpdateModules only in ProcessModulesTasks and Wait()

* clean code; update RingListener and RingModule::Process

* move EloqModule related symbols to eloq_module.h

* add comment for EloqModule

* NotifyWorker return worker running status
